updated new classes in layers branch for namespace rename from view:: -> ui::
