+++ src/ui/Graph.h
@@ -21,33 +21,55 @@
 
 #pragma once
 
+#include "ui/Renderer.h"
+#include "ui/View.h"
+
 #include "cinder/Cinder.h"
-#include "cinder/Color.h"
-#include "cinder/Rect.h"
+#include "cinder/Exception.h"
+#include "cinder/Signals.h"
 
-namespace view {
+namespace cinder { namespace app {
 
-typedef std::shared_ptr<class Renderer>     RendererRef;
+typedef std::shared_ptr<class Window>   WindowRef;
 
-class Renderer {
+} } // namespace ci::app
+
+namespace ui {
+
+typedef std::shared_ptr<class Graph>	GraphRef;
+typedef std::shared_ptr<class View>		ViewRef;
+
+//! This is where it all starts! Construct a Graph as the root of your UI scene graph, add other views to it.
+class Graph : public View {
   public:
-	//! Sets the current color used for rendering
-	void setColor( const ci::ColorA &color );
-	//! Stores the current color.
-	void pushColor();
-	//! Makes \a color the current color used when rendering, first storing the current color.
-	void pushColor( const ci::ColorA &color );
-	//! Restores the color to what was previously set before the last pushColor().
-	void popColor();
-	//! Draws a solid rectangle with dimensions \a rect.
-	void drawSolidRect( const ci::Rectf &rect );
-	//! Draws a stroked rectangle with dimensions \a rect.
-	void drawStrokedRect( const ci::Rectf &rect );
-	//! Draws a stroked rectangle centered around \a rect, with a line width of \a lineWidth
-	void drawStrokedRect( const ci::Rectf &rect, float lineWidth );
+	Graph( const ci::app::WindowRef &window = nullptr );
+	~Graph();
+
+	RendererRef getRenderer()       { return mRenderer; }
+	RendererRef getRenderer() const { return mRenderer; }
+
+	void propagateUpdate();
+	void propagateDraw();
+
+	//! Connects this View's touches propagation methods to the Window's touch event signals
+	void connectTouchEvents( int prioririty = 1 );
+	//! Disconnects touches propagation methods.
+	void disconnectEvents();
 
   private:
-	std::vector<ci::ColorA>		mColorStack;
+	RendererRef         mRenderer;
+	ci::app::WindowRef  mWindow;
+	bool                mMultiTouchEnabled = false;
+	int					mEventSlotPriority = 1;
+
+	std::vector<ci::signals::Connection>	mEventConnections;
 };
 
-} // namespace view
+class GraphExc : public ci::Exception {
+  public:
+	GraphExc( const std::string &description )
+		: Exception( description )
+	{}
+};
+
+} // namespace ui
